requirements.txt
Django==1.7
amqp==1.4.6
anyjson==0.3.3
billiard==3.3.0.18
celery==3.1.14
django-debug-toolbar==1.2.1
flake8==2.2.3
gnureadline==6.3.3
ipdb==0.8
ipython==2.2.0
kombu==3.0.22
mccabe==0.2.1
pep8==1.5.7
psycopg2==2.5.4
pyflakes==0.8.1
pytz==2014.7
sqlparse==0.1.11
wsgiref==0.1.2
